Fix test organization
